requirements.txt
click
paramiko
pyyaml
dictop
appserver
